Relocate tests for deprecated APIs
The deprecated methods have similarly named stable versions. This has
the potential to confuse future contributors, who may mistakenly modify
the tests for the wrong version of a given method.

Organize all tests for deprecated APIs into a dedicated file to help
avoid this kind of confusion.
